sunxi: Add CHIP's DIP support
The NextThing CHIP comes with expansion boards called DIPs.

These DIPs comes with a 1-Wire EEPROM used to enumerate and identify the
DIPs currently attached.

Once we know what is connected, we need to do various things, such a load
and apply an overlay if relevant, adjust the U-boot environment and the
kernel command line, etc.

Add support for this.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>